backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
sendgrid
apscheduler
httpx
python-jose
passlib
bcrypt
shodan
stripe
requests
cryptography
pydantic
pydantic-settings
pytz
dnspython
